# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ModelCreateRequest:
    properties:
      make_id:
        type: string
      name:
        type: string
      vehicle_type_id:
        type: string
    required:
    - make_id
    - name
    - vehicle_type_id
    type: object
  dto.ModelUpdateRequest:
    properties:
      id:
        type: string
      make_id:
        type: string
      name:
        type: string
      vehicle_type_id:
        type: string
    required:
    - id
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Motos API
  version: 0.1.2
paths:
...
  /api/models:
    get:
      consumes:
      - application/json
      description: Get a list of all active vehicle models with their make and type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all models
      tags:
      - models
    post:
      consumes:
      - application/json
      description: Create a new vehicle model with the input payload
      parameters:
      - description: Model Create Request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/dto.ModelCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a new model
      tags:
      - models
  /api/models/{id}:
    get:
      consumes:
      - application/json
      description: Get a vehicle model by ID with its make and type
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a model by ID
      tags:
      - models
    put:
      consumes:
      - application/json
      description: Update an existing vehicle model with the input payload
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      - description: Model Update Request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/dto.ModelUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an existing model
      tags:
      - models
  /api/models/{id}/status:
    patch:
      consumes:
      - application/json
      description: Soft delete a vehicle model by setting status to false
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Change status a model (soft delete)
      tags:
      - models
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"